Enhancement on Query Processor and draft statistics interface - changed hits from uint64_t to int; - QP_rule_t has a pointer to its parent : this is done to speed up lookup - implemented get_current_query_rules() to execute SAVE MYSQL QUERY RULES FROM RUNTIME (issue #93) - implemented get_stats_query_rules() to save statistics on stats_mysql_query_rules (issue #139) - created 2 new classes to easily process statistics in Query Processor : QP_rule_text and QP_rule_text_hitsonly - extended classes SQLite3_row and SQLite3_result - every __admin_refresh_interval milliseconds Admin refreshes statistics - defined table stats_mysql_query_rules - deprecated __attach_configdb_to_admindb() , replaced with __attach_db_to_admindb (used in issue #142) - to admindb are now attached both configdb and statsdb (issue #142) - drafted a way to update hits on Query Processor . This method will go away when issue #146 will be implemented - fixed crashing bug in issue #145

+++ lib/mysql_session.cpp
@@ -754,15 +754,17 @@ __exit_DSS__STATE_NOT_INITIALIZED:
 							status=WAITING_CLIENT_DATA;
 							client_myds->DSS=STATE_SLEEP;
 							if (qpo) {
+								if (qpo->cache_ttl>0) { // Fixed bug #145
+									client_myds->PSarrayOUT->copy_add(server_myds->resultset,0,server_myds->resultset->len);
+									unsigned char *aa=server_myds->resultset2buffer(false);
+									while (server_myds->resultset->len) server_myds->resultset->remove_index(server_myds->resultset->len-1,NULL);	
+									GloQC->set((unsigned char *)client_myds->query_SQL,strlen((char *)client_myds->query_SQL)+1,aa,server_myds->resultset_length,30);
+									l_free(server_myds->resultset_length,aa);
+									server_myds->resultset_length=0;
+									l_free(strlen((char *)client_myds->query_SQL)+1,client_myds->query_SQL);
+								}
 								GloQPro->delete_QP_out(qpo);
 								qpo=NULL;
-								client_myds->PSarrayOUT->copy_add(server_myds->resultset,0,server_myds->resultset->len);
-								unsigned char *aa=server_myds->resultset2buffer(false);
-								while (server_myds->resultset->len) server_myds->resultset->remove_index(server_myds->resultset->len-1,NULL);	
-								GloQC->set((unsigned char *)client_myds->query_SQL,strlen((char *)client_myds->query_SQL)+1,aa,server_myds->resultset_length,30);
-								l_free(server_myds->resultset_length,aa);
-								server_myds->resultset_length=0;
-								l_free(strlen((char *)client_myds->query_SQL)+1,client_myds->query_SQL);
 							}
 						}
 						break;
